Style code relating to the HSCP profiles
Used `styler::style_dir("hscp_level/")` which has applied a consistent code style to all files.

There should be no changes to how the code works, just to what it looks like. It should make the code nicer/easier to read and also make it easier to compare to the 'main' locality profile code as that has also been styled in the same way.