Tidy the SMR connection code
This is a minor QOL change that's probably only relevant in `1. Unscheduled Care data extraction.R`. It uses `Sys.getenv("USER")` to get the username from the system environment, so only the password is required to be entered.

I just copy/pasted to make sure it's the same code in all locations to avoid any confusion as to why one is different from another.