Simplify and speedup code to count localities
I'd rather we just do this once, but since we also need code for testing, I didn't want to get into that...

This PR ensures that every time the number of localities is counted it's done consistently by refactoring the code into a function.

It's also dramatically simpler and more efficient to just do a `sum` vs. the old code.