Remove unused objects at the end of the chapter scripts
I've gone through and checked which objects are created but not used (as far as I can tell). All these objects are now removed in a final `rm()` statement at the end of each chapter script.

This should reduce the overall memory requirements, and highlights some ways we could speed up the code by not creating objects which aren't used.