Suppress the messages about `.groups` from `dplyr::summarise`
This sets an option which causes the message about `.groups` to not be displayed. See: https://dplyr.tidyverse.org/reference/summarise.html#arguments

Whist `.groups` is a nice addition it's still labelled as 'experimental', as well as that, all of our code has been written without it and instead uses `ungroup()`. If `ungroup()` is ever deprecated or superseded, or possibly if `.groups` becomes stable, we should revisit this fix.