Change the project formatter from `{styler}` to `air` (#248)
* Change the project formatter from `{styler}` to `air`

Air is a new styler developed by Posit. It's slightly more 'aggressive' than `{styler}` was, but in principle follows the same tidyverse style guide.

The advantage of switching is that it is *much* faster - both the styling itself, but importantly, the installation of it in the actions runner. The styler action took 1-3 mins to install and run, whereas air should take ~5 seconds to do the same thing.

https://posit-dev.github.io/air/formatter.html
https://www.tidyverse.org/blog/2025/02/air/
https://yohann-data.fr/posts/airformatter/

* Update style-project.yaml

* Style code

* Rename excel_output.r to excel_output.R

* Style code