Cast the number to a character in `get_article`
This was caused by #287, previously, it didn't matter if it received a numeric or character, but the change to `startsWith` means that it needs to be a character now. This casts it in the function, so nothing external should be affected as the function still just returns `"-"`, `"an"` or `"a"` whatever  type the input is.